refactored more, now has the ability to input prototype and type ratio values
